Added dot.inputrc, setup hooks, updated .gitignore.
Thanks to Cory Jacobsen (github.com/unrolled) for the life-changing tip
about inputrc. In short, it now uses a partially typed command as input
for history search. You should change your files to include it now.
RIGHT now. Do it.

Added a setup hook for .inputrc.

Updated the .gitignore with a some vim generated files that get stashed
in the vim subfolder, and the VERSION utility file that's updated on each
account these are installed in.

sd.